+++ coderd/workspacestats/workspacestatstest/batcher.go
@@ -1,4 +1,4 @@
-package wstest
+package workspacestatstest
 
 import (
 	"sync"
